pyproject.toml
[project]
name = "router-api"
version = "0.1.0"
description = "a small python API that checks incoming MAC addresses against a google sheets and hands out proxy links"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "gspread>=6.2.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]